networking/requirements.txt
aws-cdk.aws-ec2==1.10.0
aws-cdk.core==1.10.0